refactor: Moved `FeatureThreshold` to a types constant file
