+++ src/colorizer/ColorizeCanvas.ts
@@ -29,6 +29,7 @@ import vertexShader from "./shaders/colorize.vert";
 import fragmentShader from "./shaders/colorize_RGBA8U.frag";
 import pickFragmentShader from "./shaders/cellId_RGBA8U.frag";
 import Track from "./Track";
+import { FeatureThreshold } from "../constants/types";
 
 const BACKGROUND_COLOR_DEFAULT = 0xf7f7f7;
 export const OUTLIER_COLOR_DEFAULT = 0xc0c0c0;
@@ -92,13 +93,6 @@ const getDefaultUniforms = (): ColorizeUniforms => {
     outlierDrawMode: new Uniform(DrawMode.USE_COLOR),
     outOfRangeDrawMode: new Uniform(DrawMode.USE_COLOR),
   };
-};
-
-export type FeatureThreshold = {
-  featureName: string;
-  units: string;
-  min: number;
-  max: number;
 };
 
 // TODO: Change into a component?
