+++ src/colorizer/utils/data_utils.ts
@@ -1,4 +1,4 @@
-import { FeatureThreshold } from "../ColorizeCanvas";
+import { FeatureThreshold } from "../../constants/types";
 
 /**
  * Generates a find function for a FeatureThreshold, matching on feature name and unit.
