+++ src/colorizer/utils/url_utils.ts
@@ -1,7 +1,7 @@
 // Typescript doesn't recognize RequestInit
 /* global RequestInit */
 
-import { FeatureThreshold } from "../ColorizeCanvas";
+import { FeatureThreshold } from "../../constants/types";
 import { numberToStringDecimal } from "./math_utils";
 
 const URL_PARAM_TRACK = "track";
