+++ src/constants/types.ts
@@ -1,0 +1,6 @@
+export type FeatureThreshold = {
+  featureName: string;
+  units: string;
+  min: number;
+  max: number;
+};
